Consolidate duplicated path/PDF helpers into shared leaf modules (#654)
